Adds AllAxes (#203)
* Adding AllAxes, combining sum/mean with axes versions

* Renaming files

* Updating docstrings
