+++ src/tensor_ops/impl_broadcast_reduce.rs
@@ -1,5 +1,5 @@
 use super::utils::move_tape_and_add_backward_op;
-use crate::arrays::{Axes2, Axes3, Axes4, Axis};
+use crate::arrays::{AllAxes, Axes2, Axes3, Axis};
 use crate::devices::{AddAccum, CopyAccum, DeviceReduce};
 use crate::prelude::*;
 
@@ -32,7 +32,7 @@ pub trait Broadcast<T, Axes> {
 pub trait Reduce<Axes>: Sized + Tensor<Dtype = f32> {
     /// The resulting tensor type.
     /// This can be broadcast into Self via [Broadcast].
-    type Reduced: Tensor<Tape = Self::Tape, Dtype = Self::Dtype> + Broadcast<Self, Axes>;
+    type Reduced: Broadcast<Self, Axes> + Tensor<Tape = Self::Tape, Dtype = Self::Dtype>;
     type DeviceR: DeviceReduce<Self::Array, Axes, Reduced = <Self::Reduced as HasArrayType>::Array>;
 }
 
@@ -65,12 +65,25 @@ impl<H: Tape> Broadcast<Tensor0D<H>, Axis<-1>> for Tensor0D<H> {
         self
     }
 }
+impl<H: Tape> Reduce<AllAxes> for Tensor0D<H> {
+    type Reduced = Self;
+    type DeviceR = <Self as HasDevice>::Device;
+}
+impl<H: Tape> Broadcast<Tensor0D<H>, AllAxes> for Tensor0D<H> {
+    fn broadcast(self) -> Tensor0D<H> {
+        self
+    }
+}
 
 // 0d -> Nd
 impl_broadcast_reduce!(Tensor0D<H>, Axis<-1>, Tensor1D<M, H>, {M});
 impl_broadcast_reduce!(Tensor0D<H>, Axes2<0, 1>, Tensor2D<M, N, H>, {M, N});
 impl_broadcast_reduce!(Tensor0D<H>, Axes3<0, 1, 2>, Tensor3D<M, N, O, H>, {M, N, O});
 impl_broadcast_reduce!(Tensor0D<H>, Axes4<0, 1, 2, 3>, Tensor4D<M, N, O, P, H>, {M, N, O, P});
+impl_broadcast_reduce!(Tensor0D<H>, AllAxes, Tensor1D<M, H>, {M});
+impl_broadcast_reduce!(Tensor0D<H>, AllAxes, Tensor2D<M, N, H>, {M, N});
+impl_broadcast_reduce!(Tensor0D<H>, AllAxes, Tensor3D<M, N, O, H>, {M, N, O});
+impl_broadcast_reduce!(Tensor0D<H>, AllAxes, Tensor4D<M, N, O, P, H>, {M, N, O, P});
 
 // 1d -> Nd
 impl_broadcast_reduce!(Tensor1D<M, H>, Axis<-1>, Tensor2D<M, N, H>, {M, N});
@@ -108,7 +121,8 @@ mod tests {
 
     #[test]
     fn test_valid_1d_broadcasts() {
-        let _: Tensor1D<5> = Tensor0D::zeros().broadcast();
+        let _: Tensor1D<5> = Broadcast::<_, AllAxes>::broadcast(Tensor0D::zeros());
+        let _: Tensor1D<5> = Broadcast::<_, Axis<-1>>::broadcast(Tensor0D::zeros());
 
         let _: Tensor2D<5, 3> = Tensor1D::<3>::zeros().broadcast();
         let _: Tensor2D<5, 3> = Tensor1D::<5>::zeros().broadcast();
@@ -126,7 +140,7 @@ mod tests {
 
     #[test]
     fn test_valid_2d_broadcasts() {
-        let _: Tensor2D<5, 3> = Tensor0D::zeros().broadcast();
+        let _: Tensor2D<5, 3> = Broadcast::<_, AllAxes>::broadcast(Tensor0D::zeros());
 
         let _: Tensor3D<3, 5, 7> = Tensor1D::<3>::zeros().broadcast();
         let _: Tensor3D<3, 5, 7> = Tensor1D::<5>::zeros().broadcast();
@@ -142,7 +156,7 @@ mod tests {
 
     #[test]
     fn test_valid_3d_broadcasts() {
-        let _: Tensor3D<3, 5, 7> = Tensor0D::zeros().broadcast();
+        let _: Tensor3D<3, 5, 7> = Broadcast::<_, AllAxes>::broadcast(Tensor0D::zeros());
 
         let _: Tensor4D<3, 5, 7, 9> = Tensor1D::<3>::zeros().broadcast();
         let _: Tensor4D<3, 5, 7, 9> = Tensor1D::<5>::zeros().broadcast();
