+++ src/tensor_ops/mod.rs
@@ -135,21 +135,19 @@ mod impl_clamp;
 mod impl_div;
 mod impl_dropout;
 mod impl_mask;
-mod impl_max_axes;
+mod impl_max;
 mod impl_maximum;
 mod impl_mean;
-mod impl_mean_axes;
-mod impl_min_axes;
+mod impl_min;
 mod impl_minimum;
 mod impl_mul;
 mod impl_nans;
-mod impl_normalize_axes;
+mod impl_normalize;
 mod impl_pow;
 mod impl_softmax;
-mod impl_std_axes;
+mod impl_std;
 mod impl_sub;
 mod impl_sum;
-mod impl_sum_axes;
 mod map;
 mod matmul;
 mod permute;
@@ -164,21 +162,19 @@ pub use impl_clamp::*;
 pub use impl_div::*;
 pub use impl_dropout::*;
 pub use impl_mask::*;
-pub use impl_max_axes::*;
+pub use impl_max::*;
 pub use impl_maximum::*;
 pub use impl_mean::*;
-pub use impl_mean_axes::*;
-pub use impl_min_axes::*;
+pub use impl_min::*;
 pub use impl_minimum::*;
 pub use impl_mul::*;
 pub use impl_nans::*;
-pub use impl_normalize_axes::*;
+pub use impl_normalize::*;
 pub use impl_pow::*;
 pub use impl_softmax::*;
-pub use impl_std_axes::*;
+pub use impl_std::*;
 pub use impl_sub::*;
 pub use impl_sum::*;
-pub use impl_sum_axes::*;
 pub use map::*;
 pub use matmul::*;
 pub use permute::{Permute2D, Permute3D, Permute4D};
